Spanish-Surname-Separation
This script uses the Nth-character function to separate most common surnames and second names encountered in Spanish names. It is bound to not be perfect because of names too long to separate or uncommon names.

It also depends on the name's initial to be only 1 character in length, like for example, Jesus D De La Victoria Del Campo would be separated to: Jesus, D, De La Victoria, Del Campo. Jesus David De La Victoria Del Campo would encounter problems because we have no way of knowing if the second word in the string is a last name or a second name.